Organization the Project and Adding Basic Navigation
