+++ lib/Components/app_shared_text_rotules.dart
@@ -1,0 +1,2 @@
+
+const String appTitle = 'Segurança Social + Próxima';
